+++ ui/src/dashboards/components/NoteEditor.tsx
@@ -1,18 +1,15 @@
 // Libraries
 import React, {PureComponent} from 'react'
 import {connect} from 'react-redux'
-import classnames from 'classnames'
 
 // Components
 import {
   Stack,
   Alignment,
-  ButtonShape,
   SlideToggle,
   ComponentSize,
   ComponentSpacer,
 } from '@influxdata/clockface'
-import {Radio} from 'src/clockface'
 import NoteEditorText from 'src/dashboards/components/NoteEditorText'
 import NoteEditorPreview from 'src/dashboards/components/NoteEditorPreview'
 
@@ -31,7 +28,6 @@ import {AppState} from 'src/types/v2'
 
 interface StateProps {
   note: string
-  isPreviewing: boolean
   toggleVisible: boolean
   showNoteWhenEmpty: boolean
 }
@@ -48,60 +44,28 @@ type Props = StateProps & DispatchProps & OwnProps
 
 class NoteEditor extends PureComponent<Props> {
   public render() {
-    const {note, isPreviewing, onSetIsPreviewing, onSetNote} = this.props
+    const {note, onSetNote} = this.props
 
     return (
       <div className="note-editor">
-        <div className={this.controlsClassName}>
-          <div className="note-editor--radio">
-            <Radio shape={ButtonShape.StretchToFit}>
-              <Radio.Button
-                value={false}
-                active={!isPreviewing}
-                onClick={onSetIsPreviewing}
-              >
-                Compose
-              </Radio.Button>
-              <Radio.Button
-                value={true}
-                active={isPreviewing}
-                onClick={onSetIsPreviewing}
-              >
-                Preview
-              </Radio.Button>
-            </Radio>
+        <div className="note-editor--controls">
+          <div className="note-editor--helper">
+            Need help using Markdown? Check out{' '}
+            <a
+              href="https://daringfireball.net/projects/markdown/syntax"
+              target="_blank"
+            >
+              this handy guide
+            </a>
           </div>
           {this.visibilityToggle}
         </div>
         <div className="note-editor--body">
-          {this.noteEditorPreview}
           <NoteEditorText note={note} onChangeNote={onSetNote} />
-        </div>
-        <div className="note-editor--footer">
-          Need help using Markdown? Check out{' '}
-          <a
-            href="https://daringfireball.net/projects/markdown/syntax"
-            target="_blank"
-          >
-            this handy guide
-          </a>
+          <NoteEditorPreview note={note} />
         </div>
       </div>
     )
-  }
-
-  private get controlsClassName(): string {
-    const {toggleVisible} = this.props
-
-    return classnames('note-editor--controls', {centered: !toggleVisible})
-  }
-
-  private get noteEditorPreview(): JSX.Element {
-    const {isPreviewing, note} = this.props
-
-    if (isPreviewing) {
-      return <NoteEditorPreview note={note} />
-    }
   }
 
   private get visibilityToggle(): JSX.Element {